Add additional expected result for output test
The timing of the output from the xCORE and
checker has change when a pin map is specified,
causing the gpio_sim_tests::
output_test[supply_pin_map=True;timestamps=False]
test case to start failing.

The additional expected result file added is now
used to validate the result of the output test
when a pin map is supplied.